pyproject.toml
[tool.poetry]
name = "porunga"
version = "0.1.0"
description = ""
authors = ["HARIHARA SUDHAN (HARI) <108145682+hari0205@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
python-dotenv = "^1.0.1"
defusedxml = "^0.7.1"
inquirerpy = "^0.3.4"
keyring = "^25.2.1"

[tool.poetry.scripts]
porunga = "porunga.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
